+++ cde/programs/dtinfo/DtMmdb/index/dyn_disk_index.h
@@ -65,7 +65,7 @@ public:
 //init run time data components
    Boolean init_data_member( inv_lists_handler*, abs_storage* );
 
-   int handler_to_inv_idx(const handler&);
+   size_t handler_to_inv_idx(const handler&);
 
    MMDB_SIGNATURES(dyn_disk_index);
 };
